refactor: 필요없는 폴더 및 파일 삭제
