--- src/components/Common/LabeledInput/LabeledInput.style.ts
@@ -1,23 +1,0 @@
-import styled from '@emotion/styled';
-
-export const LabeledInputWrapper = styled.div`
-    display: flex;
-    align-items: center;
-`
-
-export const RecevierSpan = styled.span`
-    width: 80px;
-    font-size: 16px;
-`
-
-export const InputWrapper = styled.div`
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-`
-
-export const ReceiverWarn = styled.div`
-    font-size: 12px;
-    margin: 0 20px 5px 5px;
-    color: ${({ theme }) => theme.colors.red700};
-`
