--- src/types/orderForm.ts
@@ -1,8 +1,0 @@
-export interface TextFieldState {
-  text: string;
-  check: boolean;
-}
-
-export interface PhoneFiledState extends TextFieldState {
-  checkPhoneForm: boolean;
-}
